refactor: decompor useAppStore em stores especializadas
Decompõe a store central `app.ts` em cinco novas stores menores para melhorar a manutenabilidade, facilitar a futura validação de dados com Zod e isolar domínios de negócio:

- PacienteStore: CRUD e busca de pacientes.
- AgendamentoStore: Gestão de horários, status e remarcações.
- ProtocoloStore: Catálogo de protocolos médicos.
- PrescricaoStore: Histórico e novas prescrições.
- ConfiguracaoStore: Parâmetros globais e configurações de interface.

O arquivo `app.ts` foi mantido como uma Facade (ponto central de acesso) que integra e reexporta o estado e as ações das novas stores. Esta abordagem foi escolhida para evitar a quebra de compatibilidade com os componentes existentes, permitindo uma migração gradual das referências nos arquivos .vue sem a necessidade de refatorar todo o projeto em um único passo.
