+++ frontend/src/stores/paciente.ts
@@ -1,0 +1,85 @@
+import {defineStore} from 'pinia'
+import {ref} from 'vue'
+import api from '@/services/api'
+import {toast} from 'vue-sonner'
+import type {Paciente} from '@/types'
+
+export const usePacienteStore = defineStore('paciente', () => {
+  const pacientes = ref<Paciente[]>([])
+  const totalPacientes = ref(0)
+  const resultadosBusca = ref<Paciente[]>([])
+
+  function getPacienteById(id: string) {
+    const p = pacientes.value.find(p => p.id === id)
+    if (p && !p.idade && p.dataNascimento) {
+      const nascimento = new Date(p.dataNascimento)
+      const hoje = new Date()
+      let idade = hoje.getFullYear() - nascimento.getFullYear()
+      const m = hoje.getMonth() - nascimento.getMonth()
+      if (m < 0 || (m === 0 && hoje.getDate() < nascimento.getDate())) {
+        idade--
+      }
+      p.idade = idade
+    }
+    return p
+  }
+
+  async function fetchPacientes(page = 1, size = 10, termo = '') {
+    try {
+      const params: any = {page, size}
+      if (termo) params.termo = termo
+
+      const res = await api.get('/api/pacientes', {params})
+
+      pacientes.value = res.data.items
+      totalPacientes.value = res.data.total
+    } catch (e) {
+      console.error(e)
+    }
+  }
+
+  async function buscarPacientesDropdown(termo: string) {
+    if (!termo || termo.length < 2) {
+      resultadosBusca.value = []
+      return
+    }
+    try {
+      const res = await api.get('/api/pacientes', {params: {termo, page: 1, size: 20}})
+      resultadosBusca.value = res.data.items
+    } catch (e) {
+      console.error(e)
+    }
+  }
+
+  async function adicionarPaciente(paciente: any) {
+    try {
+      const res = await api.post('/api/pacientes', paciente)
+      pacientes.value.push(res.data)
+      return res.data
+    } catch (e) {
+      throw e
+    }
+  }
+
+  async function atualizarPaciente(id: string, dados: any) {
+    try {
+      const res = await api.put(`/api/pacientes/${id}`, dados)
+      const idx = pacientes.value.findIndex(p => p.id === id)
+      if (idx !== -1) pacientes.value[idx] = res.data
+      toast.success("Dados do paciente atualizados")
+    } catch (e) {
+      toast.error("Erro ao atualizar paciente")
+    }
+  }
+
+  return {
+    pacientes,
+    totalPacientes,
+    resultadosBusca,
+    getPacienteById,
+    fetchPacientes,
+    buscarPacientesDropdown,
+    adicionarPaciente,
+    atualizarPaciente
+  }
+})
